simplify lineBreak.needs(After/Before) & fix consecutive FunctionDeclaration.
avoid using LOC info since it seems that it is still wrong and is not really
necessary, we can simply check if previous token is a LineBreak (only way to be
on a different line is having a line break in between).

also reorder some of the config options and add more tests fo
FunctionDeclaration.