refactor utils and index. add editorconfig file.
centralize options and validate if it's a valid preset. also reduce lookups.

change the way functions are exported on utils since it's getting hard to
track, make sure we use the same names internally and externally (and keep them
next to each other).

rename and move index.js to the lib folder to improve organization.