improve whiteSpace & lineBreak needsBefore & needsAfter
it's too common to need to add a space before/after a token based on a config
option that is on the opposite group (eg. you need a space "before" token but
setting is defined on "after") so we add an option to pass a boolean value to
`needsBefore` to bypass the type check (boolean value is used instead).