better error messages and avoids terminating the process on first error. fixes #303
when formatting multiple files at once (specially with the -i flag) it should
try to format as many files as possible and output error messages for each
invalid file.

we also display each error on a single line and starts message with "Error:" to
make it easier to parse the output with other tools (eg. grep)